Use Windows types LONG instead of Linux types long etc.; patch from Frank Zago.
The offsets mentioned in comments in x86_64_stubs.S are hexadecimal; make all
offsets hexadecimal; it is more efficient to xor al register than to
move 0 to %rax, since variadic functions where it is used use only al.

+++ trunk/ndiswrapper/driver/usb.c
@@ -782,8 +782,8 @@ WRAP_EXPORT_MAP("_USBD_CreateConfigurationRequestEx@8",	USBD_CreateConfiguration
 STDCALL struct usb_interface_descriptor *
 	WRAP_EXPORT(USBD_ParseConfigurationDescriptorEx)
 	(struct usb_config_descriptor *config,
-	 void *startPos, __s32 intfNum, __s32 altSet,
-	 __s32 intfClass, __s32 intfSubClass, __s32 intfProto)
+	 void *startPos, LONG intfNum, LONG altSet,
+	 LONG intfClass, LONG intfSubClass, LONG intfProto)
 {
 	int size = config->wTotalLength;
 	char *pos = startPos;
